Moved SecurityMockMvcLoginRequestBuilder into its own class to avoid groovy compilation errors.
